requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==5.2.1
google==3.0.0
google-auth==2.40.3
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
